pyproject.toml
[project]
name = "autoresearch-aw"
version = "0.1.0"
description = "Run autoresearch from anywhere — Mac, GCP, AWS, or Azure"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "torch>=2.6.0",
    "numpy>=2.0",
    "pyarrow>=15.0",
    "rustbpe>=0.1.0",
    "tiktoken>=0.11.0",
    "requests>=2.32.0",
]

[project.scripts]
autoresearch-aw = "autoresearch_aw.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
